refactor: move review component file location
